move `Source` to the core
