+++ martin/src/cog/image.rs
@@ -2,11 +2,10 @@ use std::fs::File;
 use std::io::BufWriter;
 use std::path::Path;
 
-use martin_tile_utils::TileCoord;
+use martin_tile_utils::{TileCoord, TileData};
 use tiff::decoder::{Decoder, DecodingResult};
 
 use super::CogError;
-use crate::TileData;
 
 /// Image represents a single image in a COG file. A tiff file may contain many images.
 /// This struct contains information and methods for taking tiles from the image.
