+++ martin/src/config/file/cog.rs
@@ -4,9 +4,10 @@ use std::path::PathBuf;
 use serde::{Deserialize, Serialize};
 use url::Url;
 
+use crate::MartinResult;
 use crate::cog::CogSource;
 use crate::config::file::{ConfigExtras, SourceConfigExtras, UnrecognizedKeys, UnrecognizedValues};
-use crate::{MartinResult, TileInfoSource};
+use martin_core::tiles::BoxedSource;
 
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
 pub struct CogConfig {
@@ -25,12 +26,12 @@ impl SourceConfigExtras for CogConfig {
         false
     }
 
-    async fn new_sources(&self, id: String, path: PathBuf) -> MartinResult<TileInfoSource> {
+    async fn new_sources(&self, id: String, path: PathBuf) -> MartinResult<BoxedSource> {
         let cog = CogSource::new(id, path)?;
         Ok(Box::new(cog))
     }
 
-    async fn new_sources_url(&self, _id: String, _url: Url) -> MartinResult<TileInfoSource> {
+    async fn new_sources_url(&self, _id: String, _url: Url) -> MartinResult<BoxedSource> {
         unreachable!()
     }
 }
