+++ martin/src/config/file/file_config.rs
@@ -11,9 +11,10 @@ use url::Url;
 use crate::config::file::ConfigFileError::{
     InvalidFilePath, InvalidSourceFilePath, InvalidSourceUrl, IoError,
 };
-use crate::source::TileInfoSource;
+
 use crate::utils::{IdResolver, OptMainCache};
 use crate::{MartinError, MartinResult};
+use martin_core::tiles::BoxedSource;
 
 pub type ConfigFileResult<T> = Result<T, ConfigFileError>;
 
@@ -59,23 +60,23 @@ pub trait SourceConfigExtras: ConfigExtras {
     #[must_use]
     fn parse_urls() -> bool;
 
-    /// Asynchronously creates a new `TileInfoSource` from a **local** file `path` using the given `id`.
+    /// Asynchronously creates a new `BoxedSource` from a **local** file `path` using the given `id`.
     ///
     /// This function is called for each discovered file path that is not a URL.
     fn new_sources(
         &self,
         id: String,
         path: PathBuf,
-    ) -> impl Future<Output = MartinResult<TileInfoSource>> + Send;
+    ) -> impl Future<Output = MartinResult<BoxedSource>> + Send;
 
-    /// Asynchronously creates a new `TileInfoSource` from a **remote** `url` using the given `id`.
+    /// Asynchronously creates a new `BoxedSource` from a **remote** `url` using the given `id`.
     ///
     /// This function is called for each discovered source path that is a valid URL.
     fn new_sources_url(
         &self,
         id: String,
         url: Url,
-    ) -> impl Future<Output = MartinResult<TileInfoSource>> + Send;
+    ) -> impl Future<Output = MartinResult<BoxedSource>> + Send;
 }
 
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
@@ -231,7 +232,7 @@ pub async fn resolve_files<T: SourceConfigExtras>(
     idr: &IdResolver,
     cache: OptMainCache,
     extension: &[&str],
-) -> MartinResult<Vec<TileInfoSource>> {
+) -> MartinResult<Vec<BoxedSource>> {
     resolve_int(config, idr, cache, extension).await
 }
 
@@ -240,7 +241,7 @@ async fn resolve_int<T: SourceConfigExtras>(
     idr: &IdResolver,
     cache: OptMainCache,
     extension: &[&str],
-) -> MartinResult<Vec<TileInfoSource>> {
+) -> MartinResult<Vec<BoxedSource>> {
     let Some(cfg) = config.extract_file_config(cache)? else {
         return Ok(Vec::new());
     };
