+++ martin/src/config/file/mbtiles.rs
@@ -7,7 +7,8 @@ use url::Url;
 use crate::MartinResult;
 use crate::config::file::{ConfigExtras, SourceConfigExtras, UnrecognizedKeys, UnrecognizedValues};
 use crate::mbtiles::MbtSource;
-use crate::source::TileInfoSource;
+
+use martin_core::tiles::BoxedSource;
 
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
 pub struct MbtConfig {
@@ -25,11 +26,11 @@ impl SourceConfigExtras for MbtConfig {
     fn parse_urls() -> bool {
         false
     }
-    async fn new_sources(&self, id: String, path: PathBuf) -> MartinResult<TileInfoSource> {
+    async fn new_sources(&self, id: String, path: PathBuf) -> MartinResult<BoxedSource> {
         Ok(Box::new(MbtSource::new(id, path).await?))
     }
 
-    async fn new_sources_url(&self, _id: String, _url: Url) -> MartinResult<TileInfoSource> {
+    async fn new_sources_url(&self, _id: String, _url: Url) -> MartinResult<BoxedSource> {
         unreachable!()
     }
 }
