+++ martin/src/srv/server.rs
@@ -233,11 +233,10 @@ pub fn new_server(config: SrvConfig, state: ServerState) -> MartinResult<(Server
 #[cfg(test)]
 pub mod tests {
     use async_trait::async_trait;
-    use martin_tile_utils::{Encoding, Format, TileCoord, TileInfo};
+    use martin_tile_utils::{Encoding, Format, TileCoord, TileData, TileInfo};
     use tilejson::TileJSON;
 
-    use crate::UrlQuery;
-    use crate::source::{Source, TileData, TileInfoSource};
+    use martin_core::tiles::{BoxedSource, Source, UrlQuery};
 
     #[derive(Debug, Clone)]
     pub struct TestSource {
@@ -260,7 +259,7 @@ pub mod tests {
             TileInfo::new(Format::Mvt, Encoding::Uncompressed)
         }
 
-        fn clone_source(&self) -> TileInfoSource {
+        fn clone_source(&self) -> BoxedSource {
             Box::new(self.clone())
         }
